+++ lua/plugins/todo.lua
@@ -1,0 +1,6 @@
+return {
+  'folke/todo-comments.nvim',
+  event = 'VimEnter',
+  dependencies = { 'nvim-lua/plenary.nvim' },
+  opts = { signs = false },
+}
