+++ lua/plugins/vim-sleuth.lua
@@ -1,0 +1,3 @@
+return {
+  'tpope/vim-sleuth', -- Detect tabstop and shiftwidth automatically
+}
